Float64

+++ spline_rl/policy/bsmp_unstructured_policy.py
@@ -28,6 +28,6 @@ class BSMPUnstructuredPolicy(BSMPPolicy):
         q_b = q_0 * (1 - s) + q_d * s
         q_cps = torch.cat(q_begin[:self._n_pts_fixed_begin] + [q_b + middle_trainable_q_pts] + q_end[::-1], axis=-2)
 
-        q, q_dot, q_ddot, t, dt, duration = self.compute_trajectory(q_cps.to(torch.float32), trainable_t_cps.to(torch.float32), differentiable=True)
+        q, q_dot, q_ddot, t, dt, duration = self.compute_trajectory(q_cps, trainable_t_cps, differentiable=True)
         self._traj_no += 1
         return q, q_dot, q_ddot, t, dt, duration
